backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.2.2
numpy==1.24.3
python-dateutil==2.8.2
requests==2.31.0
faiss-cpu==1.7.4
pydantic==2.5.0